+++ Win32Emu.Tests.User32/Gdi32Tests.cs
@@ -19,7 +19,7 @@ public class Gdi32Tests : IDisposable
     public void GetStockObject_WhiteBrush_ShouldReturnValidHandle()
     {
         // Act
-        var handle = _testEnv.CallGdi32Api("GETSTOCKOBJECT", NativeTypes.StockObject.WHITE_BRUSH);
+        var handle = _testEnv.CallGdi32Api("GETSTOCKOBJECT", (int)NativeTypes.StockObject.WHITE_BRUSH);
 
         // Assert
         Assert.NotEqual(0u, handle);
@@ -29,7 +29,7 @@ public class Gdi32Tests : IDisposable
     public void GetStockObject_BlackBrush_ShouldReturnValidHandle()
     {
         // Act
-        var handle = _testEnv.CallGdi32Api("GETSTOCKOBJECT", NativeTypes.StockObject.BLACK_BRUSH);
+        var handle = _testEnv.CallGdi32Api("GETSTOCKOBJECT", (int)NativeTypes.StockObject.BLACK_BRUSH);
 
         // Assert
         Assert.NotEqual(0u, handle);
@@ -39,7 +39,7 @@ public class Gdi32Tests : IDisposable
     public void GetStockObject_DefaultGuiFont_ShouldReturnValidHandle()
     {
         // Act
-        var handle = _testEnv.CallGdi32Api("GETSTOCKOBJECT", NativeTypes.StockObject.DEFAULT_GUI_FONT);
+        var handle = _testEnv.CallGdi32Api("GETSTOCKOBJECT", (int)NativeTypes.StockObject.DEFAULT_GUI_FONT);
 
         // Assert
         Assert.NotEqual(0u, handle);
@@ -49,7 +49,7 @@ public class Gdi32Tests : IDisposable
     public void GetStockObject_SystemFont_ShouldReturnValidHandle()
     {
         // Act
-        var handle = _testEnv.CallGdi32Api("GETSTOCKOBJECT", NativeTypes.StockObject.SYSTEM_FONT);
+        var handle = _testEnv.CallGdi32Api("GETSTOCKOBJECT", (int)NativeTypes.StockObject.SYSTEM_FONT);
 
         // Assert
         Assert.NotEqual(0u, handle);
@@ -59,7 +59,7 @@ public class Gdi32Tests : IDisposable
     public void GetStockObject_NullBrush_ShouldReturnValidHandle()
     {
         // Act
-        var handle = _testEnv.CallGdi32Api("GETSTOCKOBJECT", NativeTypes.StockObject.NULL_BRUSH);
+        var handle = _testEnv.CallGdi32Api("GETSTOCKOBJECT", (int)NativeTypes.StockObject.NULL_BRUSH);
 
         // Assert
         Assert.NotEqual(0u, handle);
@@ -69,8 +69,8 @@ public class Gdi32Tests : IDisposable
     public void GetStockObject_CalledTwice_ShouldReturnSameHandle()
     {
         // Act
-        var handle1 = _testEnv.CallGdi32Api("GETSTOCKOBJECT", NativeTypes.StockObject.DEFAULT_GUI_FONT);
-        var handle2 = _testEnv.CallGdi32Api("GETSTOCKOBJECT", NativeTypes.StockObject.DEFAULT_GUI_FONT);
+        var handle1 = _testEnv.CallGdi32Api("GETSTOCKOBJECT", (int)NativeTypes.StockObject.DEFAULT_GUI_FONT);
+        var handle2 = _testEnv.CallGdi32Api("GETSTOCKOBJECT", (int)NativeTypes.StockObject.DEFAULT_GUI_FONT);
 
         // Assert - same stock object should return same handle
         Assert.Equal(handle1, handle2);
@@ -80,8 +80,8 @@ public class Gdi32Tests : IDisposable
     public void GetStockObject_DifferentObjects_ShouldReturnDifferentHandles()
     {
         // Act
-        var handle1 = _testEnv.CallGdi32Api("GETSTOCKOBJECT", NativeTypes.StockObject.WHITE_BRUSH);
-        var handle2 = _testEnv.CallGdi32Api("GETSTOCKOBJECT", NativeTypes.StockObject.BLACK_BRUSH);
+        var handle1 = _testEnv.CallGdi32Api("GETSTOCKOBJECT", (int)NativeTypes.StockObject.WHITE_BRUSH);
+        var handle2 = _testEnv.CallGdi32Api("GETSTOCKOBJECT", (int)NativeTypes.StockObject.BLACK_BRUSH);
 
         // Assert - different stock objects should return different handles
         Assert.NotEqual(handle1, handle2);
